requirements.txt
cffi==1.11.5
decorator==4.3.0
Django==1.11.8
django-bootstrap3==10.0.1
django-braces==1.13.0
django-contrib-comments==1.9.0
django-debug-toolbar==1.10.1
django-threadedcomments==1.2
djangorestframework==3.8.2
docutils==0.15b0.dev0
gunicorn==19.7.1
ibmcloudenv==0.2.1
Pillow==5.3.0
jsonpath-rw==1.4.0
livereload==2.5.2
misaka==2.1.0
ply==3.11
psycopg2==2.7.5
pycparser==2.18
python-dotenv==0.9.1
pytz==2018.5
six==1.11.0
sqlparse==0.2.4
tornado==5.1
whitenoise==3.3.1
#